pyproject.toml
[project]
name = "happybarra"
version = "0.1.0"
description = "do i have enough money for this?"
authors = [
    {name = "mbalmeo@hawitsu.xyz"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "click (>=8.1.8,<9.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "st-pages (>=1.0.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "streamlit-antd-components (>=0.3.2,<0.4.0)",
    "bokeh (>=3.7.2,<4.0.0)",
    "streamlit-bokeh (>=3.7.1,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
coverage = "^7.6.12"
pytest-cov = "^6.0.0"
black = "^25.1.0"

[tool.poetry.group.dev.dependencies]
streamlit = "^1.42.2"
jupyterlab = "^4.3.5"
supabase = "^2.13.0"
ruff = "^0.11.0"

[tool.ruff.lint]
extend-select = ["I"]
extend-fixable = ["I"]